관리자 메뉴 - 회원 예약 내역/접종 완료 수정
