+++ ncv/src/main/webapp/admin/shot/shot_complete.jsp
@@ -61,7 +61,7 @@ ReservationDto reservationDto = reservationDao.editGet(resNo);
 				</div>
 				
 		<div>
-	<form action="../shot/complete.txt" method="post">
+	<form action="<%=request.getContextPath()%>/admin/shot/complete.txt" method="post">
 			<input type="hidden" name="resOkNo" value = "<%=reservationDto.getResNo()%>">
 			<input type="hidden" name="vaccineNo" value = "<%=reservationDto.getVaccineNo()%>">
 			<input type="hidden" name="clinicNo" value = "<%=reservationDto.getClinicNo()%>">
