+++ ncv/src/main/webapp/admin/shot/shot_list.jsp
